Add FilterErr and MapErr slice functions and refactor other tests slightly
New functions FilterErr and MapErr which do what they say on the tin but
with error handling. For those cases where it needs to be a bit more
robust. Also makes the tests use `t.Run` and switches the comparison to
`go-cmp` because `reflect.DeepEqual` has unexpected behavior! Also
go-cmp's output is way nicer.