feat: "file" provider
Much like the "hosts_file" provider, this will generate a file (or
multiple files) on either the local filesystem or a remote one. The
difference is you get to construct the contents of it from Lua! This a
little higher level than the "custom" provider but still hella flexible
and nearly negates the need for the "hosts_file" provider.