fix: disable keepalives and connection pooling
For some strange reason the built-in TCP handling stuff in Go is leaking
goroutines and leaving TCP connections open unnecessarily, which also
causes memory to leak. I'm not really sure why, but disabling all of the
pooling and keepalive stuff fixes the issue. Unfortunately this means
that HTTP requests will be slower overall, but that's preferable to
getting OOMKilled every few hours.